refactor(engine): ActionProcessor — player-action processing + effects hub
Phase 2 engine-seams PR-11 (Commit B of 2) — the LAST and highest-risk engine
extraction. Verbatim move of the action/effects domain out of GameEngine into
shared/engine/processors/ActionProcessor.ts: processAction (router),
processRoleMeeting, processExecutiveActions, processMarketing, processArtistDialogue,
applyEffects (the coupling hub every action routes through), and processDelayedEffects
(moved here from Commit A because it drains delayed flags through applyEffects).

Bodies are character-identical; only this. rebound to WeekContext. Cross-cutting calls:
- intra-class stay this.<method>(ctx, ...) (router -> handlers; role-meeting/dialogue/
  delayed-effects -> applyEffects; role-meeting -> processExecutiveActions).
- selectHighestPopularityArtist / loadSignedArtists (moved to ArtistStateProcessor in
  Commit A) are called via new ArtistStateProcessor().<method>(ctx) — PR-9 precedent.
- processMarketing's this.calculatePressPickups(...) is inlined to the exact
  FinancialSystem call the engine wrapper made (ctx.financialSystem.calculatePressPickups
  with () => ctx.getRandom(0,1) + getAccessChance.bind), preserving the seeded
  press-pickup roll in the same stream/position.

Engine keeps same-signature private delegates for all seven methods so the advanceWeek
pipeline (unchanged order) and the mood test suites (which call applyEffects via
(engine as any).applyEffects) keep working byte-for-byte. GameEngineAction is now
exported for the processor. game-engine.ts 2238 -> 1589 (2667 -> 1589 across PR-11;
now orchestrator + delegates + utils).

Golden master 8/8 (zero updates, x2). Gate suites green: mood-dialogue-integration +
tests/engine/ + artist-signing + sweep-migrations characterization (14 files / 200 tests).
